(gce) refactored autoscaler components (#2540)
