+++ app/scripts/modules/google/autoscalingPolicy/components/basicSettings/basicSettings.component.html
@@ -1,0 +1,52 @@
+<ng-form name="basicSettingsForm">
+  <div class="row">
+      <div class="col-md-3 sm-label-right">
+        Minimum number of instances
+      </div>
+      <div class="col-md-2 content-fields">
+        <input type="number"
+               class="form-control input-sm"
+               name="minNumReplicas"
+               ng-model="$ctrl.policy.minNumReplicas"
+               min="0"
+               max="{{ $ctrl.policy.maxNumReplicas }}"
+               required/>
+      </div>
+      <div class="col-md-4 error-message" ng-if="basicSettingsForm.minNumReplicas.$error.min">
+        Must be greater than zero.
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-md-3 sm-label-right">
+        Maximum number of instances
+      </div>
+      <div class="col-md-2 content-fields">
+        <input type="number"
+               class="form-control input-sm"
+               min="{{ $ctrl.policy.minNumReplicas }}"
+               required
+               name="maxNumReplicas"
+               ng-model="$ctrl.policy.maxNumReplicas"/>
+      </div>
+      <div class="col-md-4 error-message" ng-if="basicSettingsForm.maxNumReplicas.$error.min || basicSettingsForm.minNumReplicas.$error.max">
+        Must be greater than or equal to minimum.
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-md-3 sm-label-right">
+        Cool-down period (seconds)
+        <help-field key="gce.serverGroup.scalingPolicy.coolDownPeriodSec"></help-field>
+      </div>
+      <div class="col-md-2 content-fields">
+        <input type="number"
+               class="form-control input-sm"
+               min="15"
+               required
+               name="coolDownPeriodSec"
+               ng-model="$ctrl.policy.coolDownPeriodSec"/>
+      </div>
+      <div class="col-md-4 error-message" ng-if="basicSettingsForm.coolDownPeriodSec.$error.min">
+        Must be greater than or equal to 15 seconds.
+      </div>
+  </div>
+</ng-form>
